Moves home_screen file inside presentation
